+++ cloelib/cosmology/camb_cosmology.py
@@ -1,7 +1,8 @@
 """Implementation of Background and Perturbation cosmology using CAMB."""
 
 # cloelib imports
-from cloelib.cosmology.cosmology import Background, ensure_z_zero_included
+from cloelib.cosmology.cosmology import Background
+from cloelib.auxiliary.math_utils import ensure_z_zero_included
 
 # General imports
 import numpy as np
@@ -317,12 +318,9 @@ class CAMBLinearPerturbations:
             redshifts=self.z, kmax=self.kmax
         )
         self.results = camb.get_results(self.background.interface_args["CAMBparams"])
-
         self.k, _, self.Pk = self.results.get_linear_matter_power_spectrum(
             hubble_units=False, k_hunit=False
         )
-
-        self.sigma8_0 = self.results.get_sigma8().max()
 
     def matter_power_spectrum(
         self, zs: np.ndarray, ks: np.ndarray, hubble_units=False, k_hunit=False
@@ -401,6 +399,11 @@ class CAMBLinearPerturbations:
 
         return D_z_k
 
+    def sigma8_0(self) -> float:
+        """Retrieve sigma8 at z=0."""
+
+        return self.results.get_sigma8().max()
+
 
 class CAMBNonLinearPerturbations:
     """A wrapper for CAMB nonlinear perturbation calculations."""
@@ -459,8 +462,6 @@ class CAMBNonLinearPerturbations:
         self.k, _, self.Pk = self.results.get_nonlinear_matter_power_spectrum(
             hubble_units=False, k_hunit=False
         )
-
-        self.sigma8_0 = self.results.get_sigma8().max()
 
     def matter_power_spectrum(
         self, zs: np.ndarray, ks: np.ndarray, hubble_units=False, k_hunit=False
@@ -536,3 +537,8 @@ class CAMBNonLinearPerturbations:
             / self.matter_power_spectrum(np.array([0.0]), ks)[0]
         )
         return D_z_k
+
+    def sigma8_0(self) -> float:
+        """Retrieve sigma8 at z=0."""
+
+        return self.results.get_sigma8().max()
